# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: Swagger spec generated from the provided Postman collection.
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
security: []  # default none; per-route add as needed
paths:
...
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "401": { description: Unauthorized }
    put:
      tags: [Users]
      summary: Update current user
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              name: Ahmed Updated
              phone: "03009876543"
              city: Karachi
              age: 29
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
  /api/v1/users/me/avatar:
    post:
      tags: [Users]
      summary: Upload avatar
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200": { description: Success }
  /api/v1/users/me/bookmarks:
    post:
      tags: [Users]
      summary: Add bookmark
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              itemType: right
              itemId: 1
      responses:
        "200": { description: Success }
    get:
      tags: [Users]
      summary: List bookmarks
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Success }
  /api/v1/users/me/bookmarks/{bookmarkId}:
    delete:
      tags: [Users]
      summary: Delete bookmark
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: bookmarkId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
  /api/v1/users/me/activity:
    post:
      tags: [Users]
      summary: Log activity
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              eventType: viewed_right
              payload:
                rightId: 5
                timestamp: "2025-01-15T10:30:00Z"
      responses:
        "200": { description: Success }
    get:
      tags: [Users]
      summary: Get activity
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Success }

  # -------------------- RIGHTS --------------------
  /api/v1/rights:
    get:
      tags: [Rights]
      summary: List rights
      parameters:
        - in: query
          name: category
          required: false
          schema: { type: string }
        - in: query
          name: language
          required: false
          schema: { type: string }
      responses:
        "200": { description: Success }
    post:
      tags: [Rights]
      summary: Create right (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              topic: Right to Education
              body: Every child has the right to free education...
              category: education
              language: en
              tags: [education, children, rights]
      responses:
        "200": { description: Success }
  /api/v1/rights/{rightId}:
    get:
      tags: [Rights]
      summary: Get single right
      parameters:
        - in: path
          name: rightId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
    put:
      tags: [Rights]
      summary: Update right (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: rightId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              body: Updated body text...
              tags: [education, updated]
      responses:
        "200": { description: Success }
    delete:
      tags: [Rights]
      summary: Delete right (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: rightId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }

  # -------------------- TEMPLATES --------------------
  /api/v1/templates:
    get:
      tags: [Templates]
      summary: List templates
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: language
          schema: { type: string }
      responses:
        "200": { description: Success }
    post:
      tags: [Templates]
      summary: Create template (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: Rent Agreement
              description: Standard rent agreement template
              body: This agreement is made between {{landlordName}} and {{tenantName}}...
              category: property
              language: en
              tags: [rent, property]
      responses:
        "200": { description: Success }
  /api/v1/templates/{templateId}:
    get:
      tags: [Templates]
      summary: Get single template
      parameters:
        - in: path
          name: templateId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
    put:
      tags: [Templates]
      summary: Update template (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: templateId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: Updated Rent Agreement
              body: Updated template body...
      responses:
        "200": { description: Success }
    delete:
      tags: [Templates]
      summary: Delete template (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: templateId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }

  # -------------------- PATHWAYS --------------------
...